🎨 🚚 라벨 컴포넌트를 만들면서 중복되는 코드를 분리했어요.
